Made a new project called HelloASPDotNET and used a template with a MSV Controller. Then I started on HelloController.cs to start learning how to make some routes and the content inside them. I got a starting page which asks for you name and once you hit enter it will change the page to one that says "Welcome to my app, {name}!" OR you can navigate directly to that page via the correct route of /hello and it will have "World" in place of {name}.
I am following along a tutorial which I will add to my README before pushing (not sure how VisualStudio works yet so maybe I will need to do that in the next push)

FORGOT to say that I am also learning VisualStudio as I heard from a employer that their company used it quite a bit so I thought I would go ahead and at least familiarize myself with it. I can see why people might like it so far but I currently don't really like the style of it.

Figured out how to find a file outside of the current project context so I could update the README! Once I found it, it made sense but I was hoping that I when I made the new project I could see it from the repo as the starting point. Still learning so who knows yet.

Certain file caused an issue when trying to push, added to gitignore (also created one)