So for the final part of this tutorial we started learning about making our own views. Similar to making components in React, we basically  separate out what is actually going into the component into its own separate component. Not entirely sure how accurate this is but this is my current understanding.
So instead of having all of the information directly typed into one file, we have each action having its own view to return. This seems to allow a great deal of readability to the code and keeps the overall length of files down. It also seems very functional and seemless, to be honest.

Took out the contents from both Index and Welcome actions in our HelloController and put them into their corresponding views in the corresponding folder. Of course had to update the code a bit to ensure it would work properly but now it the html doesn't need the excess C# code containing it.

This was the final entry to this part of the tutorial HelloASPDotNET. Moving onto a new project called CodingEvents now (within the same tutorial playlist)